requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
jsonpatch==1.33
jsonpointer==2.4
langchain-core==0.1.52
langchain-openai==0.1.6
langsmith==0.1.58
moviepy==1.0.3
numpy==1.26.4
openai==1.30.1
opencv-python==4.9.0.80
orjson==3.10.3
packaging==23.2
pillow==10.3.0
proglog==0.1.10
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
sniffio==1.3.1
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
